+++ src/Structures/Event.js
@@ -13,7 +13,6 @@ function RunFunction(client, ...eventArgs) {}
  */
 class Event {
   /**
-   *
    * @param {K} event
    * @param {RunFunction<K>} runFunction
    */